apply pundit policy to bills and route non-active accounts to account edit

+++ app/controllers/bills_controller.rb
@@ -1,12 +1,14 @@
 class BillsController < ApplicationController
   before_action :authenticate_user!
-  helper_method :sort_column, :sort_direction
+  # after_action :verify_authorized
+  # after_action :verify_policy_scoped, except: [:new, :create]
+  # helper_method :sort_column, :sort_direction
 
   def index
     @bills = current_account.bills.uninvoiced.includes(:customer, :supplier, :category, :vat_rate).
                                               page(params[:page]).
                                               order(sort_column + " " + sort_direction)
-
+    authorize @bills
     respond_to do |format|
       format.html
       format.csv  { 
@@ -19,6 +21,7 @@ class BillsController < ApplicationController
 
   def new
     @bill = current_account.bills.build
+    authorize @bill
     @bill.date = Date.today
     @bill.customer_id = default_customer.id if default_customer
   end
@@ -26,6 +29,7 @@ class BillsController < ApplicationController
   def create
     collect_new_entries
     @bill = Bill.new(bill_params)
+    authorize @bill
     if @bill.save
       redirect_to bills_url, notice: 'Bill was succesfully created'
     else
@@ -35,11 +39,13 @@ class BillsController < ApplicationController
 
   def edit
     @bill = current_account.bill(params[:id])
+    authorize @bill
   end
 
   def update
     collect_new_entries
     @bill = current_account.bill(params[:id])
+    authorize @bill
     if @bill.update(bill_params)
       flash[:success] = "Bill successfully updated"
       respond_to do |format|
@@ -53,6 +59,7 @@ class BillsController < ApplicationController
 
   def destroy
     @bill = current_account.bill(params[:id])
+    authorize @bill
     if @bill.destroy
       flash[:success] = "Bill successfully deleted"
       redirect_to bills_url, status: 303
@@ -61,21 +68,30 @@ class BillsController < ApplicationController
     end
   end
 
-  def category_chart
-    render json: current_account.bills.
-                      includes(:category).
-                      group("categories.name").
-                      references(:category).
-                      order("categories.name").
-                      sum(:amount)
-  end
+  # def category_chart
+  #   @bills = current_account.bills.
+  #                     includes(:category).
+  #                     group("categories.name").
+  #                     references(:category).
+  #                     order("categories.name").
+  #                     sum(:amount)
+  #   authorize @bills
+  #   render json: @bills
+  #   # render json: current_account.bills.
+  #   #                   includes(:category).
+  #   #                   group("categories.name").
+  #   #                   references(:category).
+  #   #                   order("categories.name").
+  #   #                   sum(:amount)
+  # end
 
   private
   def bill_params
-    params.require(:bill).permit(:account_id, :customer_id, :supplier_id, 
-                                 :date, :category_id, :description, :amount, 
-                                 :new_customer, :new_supplier, :new_category,
-                                 :vat_rate_id, :vat)
+    # params.require(:bill).permit(:account_id, :customer_id, :supplier_id, 
+    #                              :date, :category_id, :description, :amount, 
+    #                              :new_customer, :new_supplier, :new_category,
+    #                              :vat_rate_id, :vat)
+    params.require(:bill).permit(*policy(@bill || Bill).permitted_attributes)
   end
 
   def collect_new_entries
